feat: ajout de clients et vehicules dans le controller debug

+++ src/main/java/fr/formation/ecf/backend/ecf3cdabackend/vehicules/Vehicule.java
@@ -17,7 +17,7 @@ public class Vehicule extends EntityInfo {
     private String modele;
     private String immatriculation;
     private String etat;
-    private Double prixJournee;
+    private Integer prixJournee;
     private Boolean disponible;
     private String type;
